restructure directories on server
