+++ package.json
@@ -2,10 +2,10 @@
   "name": "f-logger",
   "version": "0.1.0",
   "description": "A friendly student logger",
-  "main": "index.js",
+  "main": "server/index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "server:watch": "nodemon index.js",
+    "server:watch": "nodemon server/index.js",
     "server": "node app.js",
     "build:watch": "./node_modules/.bin/webpack -d --watch",
     "build": "./node_modules/.bin/webpack -p",
@@ -32,6 +32,8 @@
     "mongoose": "^5.2.9",
     "mongoose-unique-validator": "^2.0.1",
     "nunjucks": "^3.1.3",
+    "passport": "^0.4.0",
+    "passport-local": "^1.0.0",
     "react": "^16.4.2",
     "react-dom": "^16.4.2",
     "react-router-dom": "^4.3.1",
